# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mera Ghar Sansar API Docs",
    "version": "1.0.0",
    "description": "Auto-generated API documentation",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development Server"
    },
    {
      "url": "https://your-production-url.com",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/api/v1/category/create": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "price": {
                    "example": "any"
                  },
                  "premiumPrice": {
                    "example": "any"
                  },
                  "premiumPlusPrice": {
                    "example": "any"
                  },
                  "autoFilled": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/category/update/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "price": {
                    "example": "any"
                  },
                  "premiumPrice": {
                    "example": "any"
                  },
                  "premiumPlusPrice": {
                    "example": "any"
                  },
                  "active": {
                    "example": "any"
                  },
                  "autoFilled": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/category/delete/{id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/category/getAll": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/category/purchase": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "example": "any"
                  },
                  "vendorId": {
                    "example": "any"
                  },
                  "categoryId": {
                    "example": "any"
                  },
                  "transactionId": {
                    "example": "any"
                  },
                  "paymentMode": {
                    "example": "any"
                  },
                  "paymentMethod": {
                    "example": "any"
                  },
                  "assignedByAdmin": {
                    "example": "any"
                  },
                  "status": {
                    "example": "any"
                  },
                  "isAdmin": {
                    "example": "any"
                  },
                  "priceTier": {
                    "example": "any"
                  },
                  "selectedPrice": {
                    "example": "any"
                  },
                  "finalPrice": {
                    "example": "any"
                  },
                  "couponCode": {
                    "example": "any"
                  },
                  "couponId": {
                    "example": "any"
                  },
                  "discountAmount": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/category/purchased/{vendorId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "vendorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/category/purchasers/{categoryId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/category/pending": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/category/pending/{vendorId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "vendorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/category/approve/{purchaseId}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "purchaseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/category/reject/{purchaseId}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "purchaseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "example": "any"
                  },
                  "reason": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/category/bulk-update-references": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "updates": {
                    "example": "any"
                  },
                  "token": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
    "/api/v1/category/create-properties-for-existing": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Category"
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}